+++ src/graph/make/faces.js
@@ -7,6 +7,9 @@ import {
 } from "../../math/polygon.js";
 import {
 	add,
+	average,
+	average2,
+	average3,
 } from "../../math/vector.js";
 import {
 	walkPlanarFaces,
@@ -34,7 +37,7 @@ import {
  * var faces = makePlanarFaces(graph);
  * faces_vertices = faces.map(el => el.vertices);
  * faces_edges = faces.map(el => el.edges);
- * @linkcode Origami ./src/graph/make.js 694
+ * @linkcode
  */
 export const makePlanarFaces = ({
 	vertices_coords, vertices_vertices, vertices_edges,
@@ -77,8 +80,9 @@ export const makePlanarFaces = ({
  * in that polygon's faces_vertices array.
  * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
  * @param {number} [epsilon=1e-6] an optional epsilon
- * @returns {number[][][]} array of array of points, where each point is an array of numbers
- * @linkcode Origami ./src/graph/make.js 820
+ * @returns {[number, number][][] | [number, number, number][][]} an array
+ * of array of points, where each point is an array of 2 or 3 numbers
+ * @linkcode
  */
 export const makeFacesPolygon = ({ vertices_coords, faces_vertices }, epsilon) => faces_vertices
 	.map(verts => verts.map(v => vertices_coords[v]))
@@ -86,39 +90,63 @@ export const makeFacesPolygon = ({ vertices_coords, faces_vertices }, epsilon) =
 
 /**
  * @description map vertices_coords onto each face's set of vertices,
- * turning each face into an array of points. "Quick" meaning collinear vertices are
- * not removed, which in some cases, this will be the preferred method.
+ * turning each face into an array of points. "Quick" meaning collinear vertices
+ * are not removed, which in some cases, this will be the preferred method.
  * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
  * @returns {number[][][]} array of array of points, where each point is an array of numbers
- * @linkcode Origami ./src/graph/make.js 831
+ * @linkcode
  */
-export const makeFacesPolygonQuick = ({ vertices_coords, faces_vertices }) => faces_vertices
-	.map(verts => verts.map(v => vertices_coords[v]));
+export const makeFacesPolygonQuick = ({ vertices_coords, faces_vertices }) => (
+	faces_vertices.map(verts => verts.map(v => vertices_coords[v]))
+);
 
 /**
  * @description For every face, get the face's centroid.
  * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
- * @returns {number[][]} array of points, where each point is an array of numbers
- * @linkcode Origami ./src/graph/make.js 839
+ * @returns {[number, number][]} array of points, where each point is an array of numbers
+ * @linkcode
  */
-export const makeFacesCenter2D = ({ vertices_coords, faces_vertices }) => faces_vertices
-	.map(fv => fv.map(v => vertices_coords[v]))
-	.map(coords => centroid(coords));
+export const makeFacesCentroid2D = ({ vertices_coords, faces_vertices }) => (
+	faces_vertices
+		.map(fv => fv.map(v => vertices_coords[v]))
+		.map(coords => centroid(coords))
+);
 
 /**
- * @description This uses point average, not centroid, faces must
- * be convex, and again it's not precise, but in many use cases
- * this is often more than sufficient.
+ * @description For every face, get the average of the face's vertices.
+ * This is not not precise, and only relevant for faces which are convex.
  * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
- * @returns {number[][]} array of points, where each point is an array of numbers
- * @linkcode Origami ./src/graph/make.js 850
+ * @returns {[number, number][]} array of points, where each point is an array of numbers
+ * @linkcode
  */
-export const makeFacesConvexCenter = ({ vertices_coords, faces_vertices }) => {
-	const oneVertex = vertices_coords.filter(() => true).shift();
-	if (!oneVertex) { return faces_vertices.map(() => []); }
-	const dimensions = oneVertex.length;
-	return faces_vertices.map(vertices => vertices
-		.map(v => vertices_coords[v])
-		.reduce((a, b) => add(a, b), Array(dimensions).fill(0))
-		.map(el => el / vertices.length));
-};
+export const makeFacesCenter2DQuick = ({ vertices_coords, faces_vertices }) => (
+	makeFacesPolygonQuick({ vertices_coords, faces_vertices })
+		.map(coords => average2(...coords))
+);
+
+/**
+ * @description For every face, get the average of the face's vertices.
+ * This is not not precise, and only relevant for faces which are convex.
+ * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
+ * @returns {[number, number, number][]} array of points, where each point is an array of numbers
+ * @linkcode
+ */
+export const makeFacesCenter3DQuick = ({ vertices_coords, faces_vertices }) => (
+	makeFacesPolygonQuick({ vertices_coords, faces_vertices })
+		.map(coords => average3(...coords))
+		// assume vertices_coords is 3D, if not, center point[2] will be NaN, fix it
+		.map(point => (Number.isNaN(point[2]) ? [point[0], point[1], 0] : point))
+);
+
+/**
+ * @description For every face, get the average of the face's vertices.
+ * This is not not precise, and only relevant for faces which are convex.
+ * @param {FOLD} graph a FOLD graph, with vertices_coords, faces_vertices
+ * @returns {[number, number][] | [number, number, number][]} array of points,
+ * where each point is an array of either 2 or 3 numbers.
+ * @linkcode
+ */
+export const makeFacesCenterQuick = ({ vertices_coords, faces_vertices }) => (
+	makeFacesPolygonQuick({ vertices_coords, faces_vertices })
+		.map(coords => average(...coords))
+);